Fury is very fast, especially when reading.
I don't have any benchmarks for anything faster right now. Wow.

+++ src/main/java/net/adoptopenjdk/bumblebench/examples/FastJsonWriteBench.java
@@ -62,7 +62,7 @@ public final class FastJsonWriteBench extends MiniBench {
 			)));
 		}
 
-		int counter = 0;
+		long counter = 0;
 		for (long i = 0; i < numLoops; i++) {
 			for (int j = 0; j < numIterationsPerLoop; j++) {
 				startTimer();
